requirements.txt
annotated-types==0.6.0
anyio==3.7.1
aplus==0.11.0
appnope==0.1.4
arrow==1.3.0
astropy==6.0.1
astropy-iers-data==0.2024.4.15.2.45.49
asttokens==2.4.1
async-generator==1.10
attrs==22.2.0
Authlib==1.2.1
beautifulsoup4==4.11.2
binaryornot==0.4.4
blake3==0.4.1
boto3==1.33.10
botocore==1.33.10
bqplot==0.12.43
branca==0.7.1
bs4==0.0.1
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.0.7
cookiecutter==1.7.3
cryptography==41.0.7
cycler==0.11.0
dask==2024.4.1
dataloader==2.0
debugpy==1.8.1
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.0
executing==2.0.1
fastapi==0.68.0
fastapi-sso==0.7.1
filelock==3.13.4
Flask==2.2.2
fonttools==4.38.0
frozendict==2.4.2
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
h5py==3.11.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
idna==3.4
importlib_metadata==7.1.0
install==1.3.5
ipydatawidgets==4.3.5
ipykernel==6.29.4
ipyleaflet==0.18.2
ipympl==0.9.4
ipython==8.23.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.11.0
ipyvuetify==1.9.4
ipywebrtc==0.6.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
lightgbm==4.3.0
llvmlite==0.42.0
locket==1.0.0
manage-fastapi==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mizani==0.11.1
mpmath==1.3.0
mysqlclient==2.2.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.24.2
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==2.2.2
parso==0.8.4
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.2.0
plotnine==0.13.4
poyo==0.5.0
progressbar2==4.4.2
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
pyerfa==2.0.1.4
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python-utils==3.8.2
pythreejs==2.4.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.0
questionary==1.10.0
requests==2.28.2
rfc3986==1.5.0
rich==13.7.1
s3transfer==0.8.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
split-folders==0.5.1
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.14.2
statsmodels==0.14.2
sympy==1.12
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.4.0
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
traittypes==0.2.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.4.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.14
uvicorn==0.29.0
uvloop==0.19.0
vaex==4.17.0
vaex-astro==0.9.3
vaex-core==4.17.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.2
vaex-ml==0.18.3
vaex-server==0.9.0
vaex-viz==0.5.4
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==2.2.2
widgetsnbextension==4.0.10
wordcloud==1.8.2.2
wsproto==1.2.0
xarray==2024.3.0
xyzservices==2024.4.0
zipp==3.18.1
